Use small number for cooldowns

+++ xiv-crafting-sim/src/effect_tracker.rs
@@ -7,15 +7,15 @@ use serde::{Serialize};
 /// Effect tracker is a key value store
 /// Data is a contiguous slice of memory, if we somehow have more than abilities than space, please just increase this and don't look back.
 #[derive(Default, Debug, Serialize, Clone)]
-pub struct EffectData([Option<(Action, i32)>; 5]);
+pub struct EffectData([Option<(Action, i8)>; 5]);
 
 
 impl EffectData {
-    pub(crate) fn get_mut(&mut self, action: Action) -> Option<&mut i32> {
+    pub(crate) fn get_mut(&mut self, action: Action) -> Option<&mut i8> {
         self.0.iter_mut().flat_map(|m| m).find(|(a, _)| *a==action).map(|(_, i)| i)
     }
 
-    pub(crate) fn get(&self, action: Action) -> Option<&(Action, i32)> {
+    pub(crate) fn get(&self, action: Action) -> Option<&(Action, i8)> {
         self.0.iter().flat_map(|m| m).find(|(a, _)| *a==action)
     }
 
@@ -27,7 +27,7 @@ impl EffectData {
     }
 
     /// Insert all values
-    pub(crate) fn insert(&mut self, action: Action, value: i32) -> bool {
+    pub(crate) fn insert(&mut self, action: Action, value: i8) -> bool {
         if let Some(v) = self.0.iter_mut().find(|f| f.is_none()) {
             *v = Some((action, value));
             true
@@ -36,7 +36,7 @@ impl EffectData {
         }
     }
 
-    pub(crate) fn iter_mut(&mut self) -> Filter<IterMut<Option<(Action, i32)>>, fn(&&mut Option<(Action, i32)>) -> bool> {
+    pub(crate) fn iter_mut(&mut self) -> Filter<IterMut<Option<(Action, i8)>>, fn(&&mut Option<(Action, i8)>) -> bool> {
         self.0.iter_mut().filter(|m| m.is_some())
     }
 }
@@ -58,7 +58,7 @@ impl FromIterator<(Action, i32)> for EffectData {
     fn from_iter<I: IntoIterator<Item=(Action, i32)>>(iter: I) -> Self {
         let mut effects = EffectData::default();
         for (a, v) in iter {
-            effects.insert(a, v);
+            effects.insert(a, v as i8);
         }
         effects
     }
